Refactor for alternate middleware
