+++ src/CUAuth/Events/CUAuthenticated.php
@@ -10,7 +10,8 @@ class CUAuthenticated
     use Dispatchable, SerializesModels;
 
     public function __construct(
-        public readonly string $userId
+        public readonly string $userId,
+        public readonly string $userLookupField = 'email',
     ) {
     }
 }
